+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ------------------
 
   * raised minimum required PHP version to 5.6.0
+  * added support for argument verification with bovigo/assert
 
 
 1.1.0 (2015-10-05)